Refactor web components and add code component   - Move elements into element folder in runtime   - Add code component for highlighting code   - Refactor and fix some bugs in editor and code components
