+++ packages/runtime/package.json
@@ -51,9 +51,11 @@
     "@codemirror/basic-setup": "^0.19.0",
     "@codemirror/highlight": "^0.19.5",
     "@codemirror/lang-cpp": "^0.19.1",
+    "@codemirror/lang-html": "^0.19.3",
     "@codemirror/lang-javascript": "^0.19.1",
     "@codemirror/lang-python": "^0.19.2",
     "@codemirror/lang-sql": "^0.19.3",
+    "@codemirror/state": "^0.19.5",
     "@runno/host": "^0.1.6",
     "@runno/terminal": "^0.1.6",
     "@wasmer/wasm-transformer": "^0.10.2",
