+++ packages/runtime/src/main.ts
@@ -1,7 +1,8 @@
-import { TerminalElement } from "./terminal";
-import { EditorElement } from "./editor";
-import { ControlsElement } from "./controls";
-import { RunElement } from "./run";
+import { TerminalElement } from "./elements/terminal";
+import { EditorElement } from "./elements/editor";
+import { ControlsElement } from "./elements/controls";
+import { RunElement } from "./elements/run";
+import { CodeElement } from "./elements/code";
 import { headlessRunCommand } from "./headless";
 import { RunnoProvider } from "./provider";
 
@@ -10,6 +11,7 @@ function defineElements() {
   customElements.define("runno-editor", EditorElement);
   customElements.define("runno-controls", ControlsElement);
   customElements.define("runno-run", RunElement);
+  customElements.define("runno-code", CodeElement);
 }
 
 export {
@@ -17,6 +19,7 @@ export {
   EditorElement,
   ControlsElement,
   RunElement,
+  CodeElement,
   RunnoProvider,
   headlessRunCommand,
   defineElements,
